requirements.txt
numpy==1.23.5
pandas==1.5.3
pytesseract
Pillow
PyMuPDF
streamlit
langchain==0.1.14
openai
py2neo
openai-whisper
qdrant-client
sentence-transformers
langchain-google-genai
google-generativeai
python-dotenv
torch>=1.13.0
ffmpeg-python